refactor(sprites): split SpriteGenerator into per-asset modules
SpriteGenerator.ts had ten self-contained `generateXSprites` functions
crammed into a single 541-line file. Move each generator to its own
module under `src/systems/sprites/` (player, tiles, token, elevator,
roomElevator, doors, particles, plants, infoBoard, infra).

SpriteGenerator.ts keeps its single public `generateSprites(scene)`
entry point and becomes a thin composition root. Pure mechanical
split, no behavior change.
